Remove `generateUsers` from global helper functions
